# Doctors/public/swagger.json
{
	"components": {
...
		"schemas": {
			"SaveUpdateResDoctor": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FullName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FullName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetDoctor": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqDoctor": {
				"properties": {
					"FullName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"FullName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqDoctor": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FullName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FullName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IDoctor": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteDoctor": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUpdateResDoctor": {
				"properties": {
					"like": {},
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"disease": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"like",
					"_id",
					"FirstName",
					"email",
					"cell",
					"password",
					"adress",
					"disease",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterDoctor": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"userName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"loginDoctor": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"like": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"doctorId": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"doctorId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUpdateResPatient": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FullName": {
						"type": "string"
					},
					"resetPasswordToken": {
						"type": "string"
					},
					"resetPasswordExpires": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "number",
						"format": "double"
					},
					"otpCode": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"disease": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FullName",
					"resetPasswordToken",
					"resetPasswordExpires",
					"email",
					"otpCode",
					"cell",
					"password",
					"adress",
					"disease",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterPatient": {
				"properties": {
					"FullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"disease": {
						"type": "string"
					},
					"doctor": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"FullName",
					"email",
					"cell",
					"password",
					"adress",
					"disease",
					"doctor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"loginPatient": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetPatient": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmailVerifyPatient": {
				"properties": {
					"resetPasswordToken": {
						"type": "string"
					},
					"resetPasswordExpires": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "number",
						"format": "double"
					},
					"otpCode": {
						"type": "string"
					}
				},
				"required": [
					"resetPasswordToken",
					"resetPasswordExpires",
					"email",
					"otpCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"forgotPassword": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyOtpCodes": {
				"properties": {
					"email": {
						"type": "string"
					},
					"otpCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"email",
					"otpCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"verifyOTPCode": {
				"properties": {
					"email": {
						"type": "string"
					},
					"otpCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"email",
					"otpCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"changePassword": {
				"properties": {
					"email": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					},
					"confirmPassword": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"newPassword",
					"confirmPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"searchPatient": {
				"properties": {
					"FullName": {
						"type": "string"
					}
				},
				"required": [
					"FullName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"searchPatientByLocation": {
				"properties": {
					"disease": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"disease",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeletePatient": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "admin_token",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "Bearer",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "swaggerapis2",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/doctor/getDoctor": {
			"post": {
				"operationId": "Getdoctor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResDoctor"
								}
							}
						}
					}
				},
				"description": "jskhd ksjhdkjsahd skhdjkshdkjs sd sdsdsds",
				"summary": "Get Doctro",
				"tags": [
					"doctor"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/savedoctor": {
			"post": {
				"operationId": "Savedoctor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResDoctor"
								}
							}
						}
					}
				},
				"summary": "Save Doctor",
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/updatedoctor": {
			"put": {
				"operationId": "UpdateDoctor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/deletedoctor": {
			"delete": {
				"operationId": "Deletdoctor",
				"responses": {
					"200": {
						"description": "Doctor deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/getDoctorlist": {
			"post": {
				"operationId": "GetdoctorList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResDoctor"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/doctor/register": {
			"post": {
				"operationId": "RegisterDoctor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/loginUser": {
			"post": {
				"operationId": "LoginEmailUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/loginDoctor"
							}
						}
					}
				}
			}
		},
		"/doctor/like": {
			"put": {
				"operationId": "FindDoctorById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/like"
							}
						}
					}
				}
			}
		},
		"/patient/patient-register": {
			"post": {
				"operationId": "RegisterPatient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterPatient"
							}
						}
					}
				}
			}
		},
		"/patient/patient-loginUser": {
			"post": {
				"operationId": "LoginEmailUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/loginPatient"
							}
						}
					}
				}
			}
		},
...
		"/patient/getPatient": {
			"post": {
				"operationId": "GetPatient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetPatient"
							}
						}
					}
				}
			}
		},
...
		"/patient/forgot-password": {
			"post": {
				"operationId": "ForgotPass",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmailVerifyPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/forgotPassword"
							}
						}
					}
				}
			}
		},
		"/patient/verify-OTPCode": {
			"post": {
				"operationId": "VerifyEamil",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyOtpCodes"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/verifyOTPCode"
							}
						}
					}
				}
			}
		},
		"/patient/change-password": {
			"post": {
				"operationId": "ChangePasswords",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/changePassword"
							}
						}
					}
				}
			}
		},
		"/patient/search": {
			"post": {
				"operationId": "SearchPatient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/searchPatient"
							}
						}
					}
				}
			}
		},
		"/patient/search-by-location": {
			"post": {
				"operationId": "SearchByLocations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/searchPatientByLocation"
							}
						}
					}
				}
			}
		},
		"/patient/delete": {
			"delete": {
				"operationId": "DeletePatients",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeletePatient"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}